remove PrintString casts

+++ TOMB4/game/health.cpp
@@ -74,7 +74,7 @@ void DrawGameInfo(long timed)
 		{
 			seconds = savegame.Level.Timer / 30;
 			sprintf(buf, "%.2d:%.2d:%.2d", seconds / 60, seconds % 60, (334 * (savegame.Level.Timer % 30)) / 100);
-			PrintString(ushort(phd_winwidth >> 1), (ushort)font_height, 0, buf, 0x8000);
+			PrintString(phd_winwidth >> 1, font_height, 0, buf, 0x8000);
 		}
 
 		if (tomb4.ammo_counter)
@@ -90,7 +90,7 @@ void DrawGameInfo(long timed)
 
 					sprintf(buf, "%i", ammo);
 					length = GetStringLength(buf, 0, &btm);
-					PrintString(ushort(LaserSight ? phd_centerx + 30 : (phd_winxmax - length - 80)), phd_winymax - btm - 70, 0, buf, 0);
+					PrintString(LaserSight ? phd_centerx + 30 : (phd_winxmax - length - 80), phd_winymax - btm - 70, 0, buf, 0);
 				}
 			}
 		}
@@ -98,7 +98,7 @@ void DrawGameInfo(long timed)
 		if (ammo_change_timer)
 		{
 			ammo_change_timer--;
-			PrintString(ushort(phd_winwidth >> 1), (ushort)font_height, 5, ammo_change_buf, 0x8000);
+			PrintString(phd_winwidth >> 1, font_height, 5, ammo_change_buf, 0x8000);
 
 			if (ammo_change_timer <= 0)
 				ammo_change_timer = 0;
